+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Changes to functions
 
 * `format_value()` gains a `zap_small` argument, to prevent scientific printing of numbers if these have more decimal places than indicated by `digits`.
+* `get_weights()` now also return `NULL` when all weights were 1.
 
 ## Bug fixes